requirements.txt
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
falcon==1.4.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pyjsend==0.2.2
pylint==2.2.2
pytest==4.2.0
python-mimeparse==1.6.0
requests==2.21.0
sentry-sdk==0.6.9
six==1.12.0
urllib3==1.24.1
wrapt==1.11.1
